services/stt/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai-whisper==20231117
ffmpeg-python==0.2.0
torch==2.2.2
torchaudio==2.2.2
numpy==1.25.2
